+++ minijinja/src/lib.rs
@@ -155,7 +155,8 @@
 //!   - `debug`: if this feature is removed some debug functionality of the engine is
 //!     removed as well.  This mainly affects the quality of error reporting.
 //!   - `deserialization`: when removed this disables deserialization support for
-//!     the [`Value`](crate::value::Value) type.
+//!     the [`Value`](crate::value::Value) type, removes the `ViaDeserialize`
+//!     type and the error type no longer implements `serde::de::Error`.
 //!
 //! There are some additional features that provide extra functionality:
 //!